chore(refractor code): Refractor Validation code - add if statement code to middleware [Delivers #163640554]
