+++ server/controllers/Offices.js
@@ -1,16 +1,7 @@
 import offices from '../models/offices';
-import validateOffice from '../middleware/validateOffice';
 
 const createOffice = (req, res) => {
   const { type, name } = req.body;
-  const { error } = validateOffice(req.body);
-  if (error) {
-    const errorMessage = error.details.map(element => element.message);
-    return res.status(400).json({
-      status: 400,
-      error: errorMessage,
-    });
-  }
   const newOffice = {
     id: offices[offices.length - 1].id + 1,
     type,
