+++ server/controllers/Party.js
@@ -1,16 +1,8 @@
-import validateParty from '../middleware/validateParty';
 import parties from '../models/parties';
 
 const createParty = (req, res) => {
   const { name } = req.body;
-  const { error } = validateParty(req.body);
-  if (error) {
-  //  const errorMessage = error.details.map(element => element.message);
-    return res.status(400).json({
-      status: 400,
-      error: error.details[0].message,
-    });
-  }
+
   const party = {
     id: parties[parties.length - 1].id + 1,
     name,
@@ -60,7 +52,6 @@ const getAParty = (req, res) => {
 
 const editAParty = (req, res) => {
   const { id } = req.params;
-  const { error } = validateParty(req.body);
   let oneParty = parties.find(party => party.id === parseInt(id, 10));
   if (isNaN(id)) {
     return res.status(400).json({
@@ -72,12 +63,6 @@ const editAParty = (req, res) => {
     return res.status(404).json({
       status: 404,
       error: 'Could not find political party',
-    });
-  }
-  if (error) {
-    return res.status(400).json({
-      status: 400,
-      error: error.details[0].message,
     });
   }
   oneParty = {
