+++ server/middleware/validateOffice.js
@@ -1,6 +1,6 @@
 import Joi from 'joi';
 
-const validateOffice = (office) => {
+const validateOffice = (req, res, next) => {
   const schema = {
     type: Joi.string().trim().min(3).required(),
     name: Joi.string().trim().min(3).required(),
@@ -12,7 +12,16 @@ const validateOffice = (office) => {
     },
     abortEarly: false,
   };
-  return Joi.validate(office, schema, options);
+  const { error } = Joi.validate(req.body, schema, options);
+
+  if (error) {
+    const errorMessage = error.details.map(element => element.message);
+    return res.status(400).json({
+      status: 400,
+      error: errorMessage,
+    });
+  }
+  next();
 };
 
 export default validateOffice;
